chore: rename rpc file name
