+++ packages/get-starknet/src/rpcs/index.ts
@@ -1,3 +1,3 @@
 export * from './switch-network';
 export * from './supported-specs';
-export * from './wallet-request-account';
+export * from './request-account';
